ScoringApi Python/requirements.txt
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.10.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
command-not-found==0.3
confection==0.1.5
cryptography==3.4.8
cymem==2.0.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
filelock==3.14.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fsspec==2024.6.1
greenlet==3.0.3
gyp==0.1
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.24.5
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==3.0.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.1.1
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
packaging==24.0
pathlib2==2.3.7.post1
pillow==10.4.0
platformdirs==4.2.2
preshed==3.0.9
pycodestyle==2.11.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.7.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==2.4.7
pyrsistent==0.20.0
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-docx==1.1.2
PyYAML==5.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
soupsieve==2.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
sympy==1.13.1
systemd-python==234
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
typer==0.12.3
typing_extensions==4.10.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
virtualenv==20.26.2
wadllib==1.3.6
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==1.0.0
